FU-185 feat: adjust photographer side layout
