+++ src/app/photographer/mypage/previous/page.tsx
@@ -1,4 +1,4 @@
-import Preparing from "@/containers/photographer/mypage/preparing";
+import Preparing from "@/containers/ui/preparing";
 
 const MyPreviousPage = () => {
   return <Preparing />;
