+++ src/containers/photographer/main/index.tsx
@@ -2,6 +2,7 @@
 
 import { useState } from "react";
 import { MainViewType } from "service-types";
+import Preparing from "@/containers/ui/preparing";
 import ReservationList from "../list";
 import Controller from "./controller";
 import { mainviewStyles } from "./main.css";
@@ -13,7 +14,7 @@ const MainView = () => {
     <div className={mainviewStyles.container}>
       <Controller view={view} setView={setView} />
       {view === "list" && <ReservationList />}
-      {view === "calender" && <div>캘린더</div>}
+      {view === "calender" && <Preparing />}
     </div>
   );
 };
