+++ src/containers/photographer/product/form/discount-field-array.tsx
@@ -1,6 +1,6 @@
 import { useFieldArray, useFormContext } from "react-hook-form";
 import { ProductFormdata } from "product-types";
-import { AddButton } from "@/components/buttons/common-buttons";
+import { CustomButton } from "@/components/buttons/common-buttons";
 import DiscountInput from "./discount-input";
 import { formStyles } from "../product.css";
 
@@ -29,7 +29,10 @@ const DiscountFieldArray = () => {
           />
         );
       })}
-      <AddButton
+      <CustomButton
+        styleType="line"
+        size="md"
+        style={{ marginTop: 20 }}
         onClick={() =>
           append({
             title: "",
