+++ src/containers/photographer/product/form/image-input.tsx
@@ -3,7 +3,10 @@ import { Image } from "product-types";
 import { Modal } from "@mantine/core";
 import { useDisclosure } from "@mantine/hooks";
 import ImageThumbnail from "@/components/images/image-thumbnail";
-import { AddButton, FinishButton } from "@/components/buttons/common-buttons";
+import {
+  CustomButton,
+  FinishButton,
+} from "@/components/buttons/common-buttons";
 import { modalStyles } from "@/containers/customer/products/products.css";
 import { getUrlFromFiles } from "@/utils/image";
 import { formStyles } from "../product.css";
@@ -107,7 +110,13 @@ const ImagesInput = ({ images, setImage }: ImageInputProps) => {
           );
         })}
       </div>
-      <AddButton onClick={() => InputRef.current?.click()} title="업로드하기" />
+      <CustomButton
+        styleType="line"
+        size="md"
+        style={{ marginTop: 20 }}
+        onClick={() => InputRef.current?.click()}
+        title="업로드하기"
+      />
       <input
         type="file"
         accept="image/*"
